Reapply "Merge pull request #11 from seohyun1001/sdj"
This reverts commit e7adfa3bc235fa42a80d844cb0f159ddf44ca893.

+++ project_DiB/src/main/java/org/zerock/project_dib/accommodation/domain/AccommodationVO.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 @Builder
 public class AccommodationVO {
 
-    private int ano;
+    private Long ano;
     private String acc_name;
     private String acc_exp;
     private String acc_loc;
